Set actual content type after decrypting download
Before this change, Pantalaimon would forward the application/octet-stream
content type in response to download requests. Even though we expect this
value for encrypted attachments, Pantalaimon stores the actual content type in
the panuploadinfo table (for any uploads executed by Pantalaimon), which we use
for the decrypted response.

As a side effect, this change also fixes the client.download usage, which broke
when the function's signature was updated in matrix-nio.