Handle OIDC logins
When a user logs into a client (through Pantalaimon) using a password,
Pantalaimon executes its own login and obtains a device ID (stored in
the accounts table), which is distinct from the user-facing client's
(stored in the devicekeys table). Even if the user logs out of the
user-facing client, the Pantalaimon device remains logged in.

Now, even though a user can log in using the OIDC IdP, before this
change, Pantalaimon would not execute its own login if no password was
supplied (such as during an OIDC login). If the user were to log out of
the user-facing client, Pantalaimon would fail to handle the deletion of
its only device, resulting in Synapse errors (i.e., "Invalid macaroon
passed.") and matrix-nio errors (e.g., "Error validating response:
'next_batch' is a required property").

This change enables Pantalaimon to handle OIDC logins by supplying the
token instead of a password.