Fix redirect loop when using multiple reverse proxies
Before this change, if Pantalaimon is deployed in front of a reverse
proxy (for Synapse) and behind an ingress, navigating to the OIDC page
will fail (ERR_TOO_MANY_REDIRECTS). This happens because Synapse will
perform its own canonical redirect. E.g., "Requested URI
https://nginx/_matrix/.../oidc-* is not canonical: redirecting to
https://public_baseurl/_matrix/.../oidc-*" from the logs.

Although the reverse proxy (nginx in this case) can set the appropriate
Host header, Pantalaimon was deleting the Host header before forwarding
the request to the reverse proxy, which was being replaced by the
aiohttp library.

This change ensures that Pantalaimon preserves the Host header for OIDC
endpoints.