Decrypt to device events only once
Pantalaimon syncs in two ways: a request by the user-facing client and
the periodic sync_forever task. Both methods may decrypt the same to
device events, which can cause BAD_MESSAGE_MAC errors. To prevent one
method from interfering with the other, this change removes the
responsibility of decrypting to device events from the user-facing
client sync.

There is a chance that the sync_forever loop iteration happens too late,
delaying the decryption of the to device events necessary for the client
to decrypt the actual messages. To resolve this, we make the sync
requested by the user-facing client wait for the sync_forever loop
iteration to finish. A side effect of this change is that login now
takes as long as the full sync timeout, which we mitigate by reducing
the value.